chore(animation): optimize files
