+++ src/packages/interactive/animation/src/web/index.ts
@@ -1,10 +1,8 @@
 import { Animation, AnimationOptions } from '../types';
 import AnimationImpl from './AnimationImpl';
 
-export function createAnimation(options?: AnimationOptions): Animation {
+function createAnimation(options?: AnimationOptions): Animation {
   return new AnimationImpl(options);
 }
 
-export default {
-  createAnimation,
-};
+export default createAnimation;
