+++ src/packages/interactive/animation/src/wechat-miniprogram/index.ts
@@ -2,12 +2,10 @@ import { Animation, AnimationOptions } from '../types';
 import { CONTAINER_NAME } from '@utils/constant';
 import { normalizeOptions } from '../common';
 
-export function createAnimation(options?: AnimationOptions): Animation {
+function createAnimation(options?: AnimationOptions): Animation {
   return wx.createAnimation(
     normalizeOptions(options, CONTAINER_NAME.WECHAT),
   );
 }
 
-export default {
-  createAnimation,
-};
+export default createAnimation;
